Refactor Jest setup for clear separation of legacy and integration tests
Changes:
- Split Jest config into two projects:
    - legacy-tests for existing test.
    - React + Redux Integration test for RTK Query + MSW test.
- Renamed mocks/ to msw-handlers/ for clarity.
- Ensured fetch polyfills and necessary setups are scoped
   to integration tests.

Why:
- Isolates older tests from new React + Redux integration tests.
- Prevents interference from different test setups.
- Enhances maintainability by making test configurations explicit.
