+++ test/mswServer.js
@@ -2,7 +2,7 @@
 // This ensures that network requests made by components are handled with predefined mock responses.
 
 import { setupServer } from 'msw/node'; // MSW's Node.js version for handling API requests in Jest.
-import { handlers } from './mocks/server_handlers'; // Import all mock API handlers from `server_handlers.js`
+import { handlers } from './msw_handlers/server_handlers'; // Import all mock API handlers from `server_handlers.js`
 
 // Creates an instance of the MSW server with the specified handlers.
 // This allows us to intercept network requests and return mock responses.
